OpBurn

+++ interfaces/kits/js/storage_manager/include/storage_manager_connect.h
@@ -78,7 +78,7 @@ public:
     int32_t Unlock(const std::string &volumeId, const std::string &pazzword);
     int32_t Decrypt(const std::string &volumeId, const std::string &pazzword);
     int32_t Eject(const std::string &volumeId);
-    int32_t GetOpticalDriveOpsProgress(const std::string &volumeId, uint64_t &progress);
+    int32_t GetOpticalDriveOpsProgress(const std::string &volumeId, uint32_t &progress);
 private:
     sptr<StorageManager::IStorageManager> storageManager_ = nullptr;
     sptr<IRemoteObject::DeathRecipient> deathRecipient_ = nullptr;
